fix(auth): 修复用户访问/users/get/1遇到异常的问题
修复了由于字段id类型设置错误导致访问具体Url遇到异常的问题
已更新相应代码来处理此情况
增加了通用的http异常处理类来统一处理异常

Fixes #1

+++ dawn-data-access/src/main/java/cn/uuxl/dawn/dal/mapper/UserMapper.java
@@ -9,5 +9,5 @@ public interface UserMapper {
     
     List<User> selectAll();
     
-    User selectById(Long id);
+    User selectById(Integer id);
 }
